requirements.txt
# Core Telegram Bot Dependencies
python-telegram-bot==20.7

# OpenAI API
openai==0.28.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Image Processing
Pillow==10.1.0

# Environment Variables
python-dotenv==1.0.0

# Web Framework for Dashboard
flask==3.0.0
flask-cors==4.0.0

# Data Processing
pandas==2.1.4

# Async Support
asyncio==3.4.3
aiohttp==3.9.1

# Security - Updated to compatible version
cryptography>=41.0.0

# Logging and Monitoring
structlog==23.2.0

# Development Tools (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# System Dependencies
setuptools>=65.0.0
wheel>=0.37.0 